Renamed test-runner to launchertester
This satisfies the linter by avoiding underscores in the package name
