+++ e2e/launchertester/default_experience_test.go
@@ -1,4 +1,4 @@
-package test_runner
+package launchertester
 
 import (
 	"context"
