+++ e2e/launchertester/test_utils.go
@@ -1,4 +1,4 @@
-package test_runner
+package launchertester
 
 import (
 	"bufio"
